# openapi.yaml
openapi: 3.0.3
info:
  title: NetVisor Suite API
  version: 1.0.0
  description: REST API for documenting, administering, and monitoring office network infrastructure.
servers:
  - url: /api/v1
paths:
...
  /rooms:
    get:
      tags: [Infrastructure]
      summary: List rooms
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of rooms }
    post:
      tags: [Infrastructure]
      summary: Create room
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                floor: { type: string, maxLength: 80 }
                purpose: { type: string, maxLength: 120 }
      responses:
        '200': { description: Created room }
        '400': { description: Validation error }
  /rooms/{id}:
    patch:
      tags: [Infrastructure]
      summary: Update room
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                floor: { type: string, maxLength: 80 }
                purpose: { type: string, maxLength: 120 }
      responses:
        '200': { description: Updated room }
        '404': { description: Room not found }
    delete:
      tags: [Infrastructure]
      summary: Delete room
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Room not found }
  /outlets:
    get:
      tags: [Infrastructure]
      summary: List outlets
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of outlets }
    post:
      tags: [Infrastructure]
      summary: Create outlet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [room_id, label]
              properties:
                room_id: { type: integer }
                label: { type: string, maxLength: 80 }
                location: { type: string, maxLength: 120 }
      responses:
        '200': { description: Created outlet }
        '400': { description: Validation error }
  /outlets/{id}:
    patch:
      tags: [Infrastructure]
      summary: Update outlet
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                room_id: { type: integer }
                label: { type: string, maxLength: 80 }
                location: { type: string, maxLength: 120 }
      responses:
        '200': { description: Updated outlet }
        '404': { description: Outlet not found }
    delete:
      tags: [Infrastructure]
      summary: Delete outlet
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Outlet not found }
  /patchpanels:
    get:
      tags: [Infrastructure]
      summary: List patch panels
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of patch panels }
    post:
      tags: [Infrastructure]
      summary: Create patch panel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 80 }
                location: { type: string, maxLength: 120 }
                ports: { type: integer, default: 24 }
      responses:
        '200': { description: Created patch panel }
        '400': { description: Validation error }
  /patchpanels/{id}:
    patch:
      tags: [Infrastructure]
      summary: Update patch panel
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 80 }
                location: { type: string, maxLength: 120 }
                ports: { type: integer }
      responses:
        '200': { description: Updated patch panel }
        '404': { description: Patch panel not found }
    delete:
      tags: [Infrastructure]
      summary: Delete patch panel
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Patch panel not found }
  /cables:
    get:
      tags: [Infrastructure]
      summary: List cable runs
      parameters:
        - in: query
          name: limit
          schema: { type: integer, maximum: 500 }
        - in: query
          name: offset
          schema: { type: integer, minimum: 0 }
      responses:
        '200': { description: Array of cables }
    post:
      tags: [Infrastructure]
      summary: Create cable run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cable_id]
              properties:
                cable_id: { type: string, maxLength: 80 }
                outlet_id: { type: integer }
                patch_panel_id: { type: integer }
                patch_port: { type: string, maxLength: 80 }
                length_m: { type: number }
                cable_type: { type: string, maxLength: 16, default: Cat6 }
                test_result: { type: string, enum: [Pass, Fail, Pending], default: Pending }
                status: { type: string, enum: [Active, Inactive], default: Active }
                notes: { type: string, maxLength: 500 }
      responses:
        '200': { description: Created cable }
        '400': { description: Validation error }
  /cables/{id}:
    patch:
      tags: [Infrastructure]
      summary: Update cable run
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                cable_id: { type: string, maxLength: 80 }
                outlet_id: { type: integer }
                patch_panel_id: { type: integer }
                patch_port: { type: string, maxLength: 80 }
                length_m: { type: number }
                cable_type: { type: string, maxLength: 16 }
                test_result: { type: string, enum: [Pass, Fail, Pending] }
                status: { type: string, enum: [Active, Inactive] }
                notes: { type: string, maxLength: 500 }
      responses:
        '200': { description: Updated cable }
        '404': { description: Cable not found }
    delete:
      tags: [Infrastructure]
      summary: Delete cable run
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Deleted }
        '404': { description: Cable not found }
...